upload in mobile app

+++ src/app/component/protect-component/customers/component/customer/accounts/assets/mutual-fund/mf-service.service.ts
@@ -488,7 +488,7 @@ export class MfServiceService {
 
   mutualFundRoundAndFormat(data, noOfPlaces: number = 0) {
     if (data) {
-     data = parseInt(data)
+     data = parseFloat(data)
       if (isNaN(data)) {
         return data;
       } else {
